requirements.txt
aiohttp==3.11.16
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
bidict==0.23.1
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
flask==2.3.3
flask-cors==4.0.0
flask-discord==0.1.69
flask-socketio==5.3.6
flask-sqlalchemy==3.1.1
frozenlist==1.4.0
greenlet==3.0.3
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
jinja2==3.1.2
markupsafe==2.1.3
multidict==6.3.2
nest-asyncio==1.5.8
oauthlib==3.2.2
py-cord==2.6.1
pyjwt==2.8.0
pymongo==4.6.0
psycopg2-binary==2.9.9
python-engineio==4.9.0
python-socketio==5.11.2
requests==2.31.0
requests-oauthlib==1.3.1
simple-websocket==1.0.0
sqlalchemy==2.0.23
typing-extensions==4.13.1
urllib3==2.1.0
werkzeug==3.0.1
wsproto==1.2.0
yarl==1.18.3
zipp==3.17.0
python-dotenv==1.0.0
google-auth==2.38.0
google-api-python-client==2.166.0
selenium==4.31.0
gspread==6.2.0
oauth2client==4.1.3
webdriver-manager==4.0.2
cryptography==44.0.2
gunicorn==23.0.0
notion-client==2.2.1
google-auth-oauthlib>=1.0.0
google-api-core>=2.11.0
sentry-sdk==1.34.0
pytz==2025.2
python-dateutil==2.9.0.post0
apscheduler==3.10.4
pyinstaller==6.1.0
